Restructure the Houston integration to keep the connection stable when encountering invalid (stale, incorrect, etc) tokens.
It's a little bit of a performance hit, but basically we try and send all the notifications, then we do a read on the socket to see if there's an error.

If there was, we bump the number of failures for the device that failed to send, and then we restart the send from the device token directly after the one that failed. The performance hit comes from the repeated process, but also having to wait to read the socket because the error doesn't come through right away, there's also a little overhead from the opening / closing of all these connections.

If a particular device fails more than the `failure_tolerance` value that's been set in the initialiser, we remove the device.