Instead of checking for production, and defaulting to development; we now check for development and default to production.
Basically the reasoning behind this is that for every build type for iOS except for debug, it's using the production gateway, so we should follow suit. Currently for a production-only backend, we have no way of forcing it back to develop to test the push notifications with the remote server. With this, we'll be able to force back to development if need be as it's checked first, but by default, all environments outside of development will use the production gateway.