Bit more restructuring.
Pushed the connection back to a class constant on the `NotifyUser::Houston` class. Forgot that the class is instantiated with each push of notifications, so having the connection memoised had to be connected each time, so it wasn't helping with the time out issues.

Doing it like this let's us keep the one connection setup, and we just use some helpers to avoid the `connection.connection` stuff.