Stop using a class ivar in ApnsChannel#deliver
Originally, we used the return value of fetch_devices to set a class
ivar @devices. We think this means that occasionally, different threads
were setting and getting this value at the same time, meaning that we
were sending notifications to the wrong devices.

For instance, thread 1 sets devices. Then, thread 2 sets devices, and
creates the new APNS and GCM classes and sends before thread 1. Thread
1 manages to create APNS and GCM _after_ thread 2 has finished, and so
uses the thread 2 values for devices.

By changing @devices to devices, we have a local variable instead of
a class instance variable, and so the devices are thread-safe.