Adds a delivery model and worker that handles delivery
This is one of the bookends for the rewrite. The major aim for it being that we don't want notifications to deal with delivery directly at all, because it can't handle tracking the delivery steps per channel, so this model basically just represents a delivery request for a notification, to a given channel. The outcome being here that a notification setup to send to APNS and GCM for example, will end up with 2 delivery objects, and that will track the sent time.