Adds an aggregator class that deals with figuring out send times
The big change here comes to the existing notifications table. The main aim with the refactor is to not have notifications deal with delivery directly, so we don't need a state or sent timestamp. We can get restructure this to _just_ have a read timestamp, and that will dictate the state the noti needs to know about.