Migrates over to using AMS 0.10.0
The use of `root` in the notification serializer was causing issues for newer apps, and it's causing issues with the tests as well because the AMS version isn't actually locked to anything.

So, for 1.0.0, we're locking to 0.10.0, and going with that, even if we have to use the :json adapter manually.